# Cache listing item partials.