--- a/Gemfile
+++ b/Gemfile
@@ -80,6 +80,7 @@ gem 'paypal-sdk-permissions',
   :ref    => 'c0240bee9f94fe6338d67b4f754e1a11ce81619a'
 gem 'paypal-sdk-merchant', '~> 1.116.0'
 gem 'airbrake', '~>4.1.0'
+gem 'cache_digests'
 
 group :staging, :production do
   gem 'newrelic_rpm', '~> 3.9.1.236'