--- a/app/views/homepage/_grid_item.haml
+++ b/app/views/homepage/_grid_item.haml
@@ -1,4 +1,4 @@
-- cache(listing) do
+- cache(["grid_item", listing, listing.author, @current_community]) do
   .home-fluid-thumbnail-grid-item
     %div
       -# Listing image