--- a/app/views/homepage/_list_item.haml
+++ b/app/views/homepage/_list_item.haml
@@ -1,52 +1,53 @@
-.home-list-item
-  - if listing.listing_images.size > 0
-    = link_to listing, :class => "home-list-image-container-desktop" do
-      = image_tag listing.listing_images.first.image.url(:small_3x2), {:alt => listed_listing_title(listing), :class => "home-list-image"}
-  - if listing.listing_images.size > 0
-    = link_to listing, :class => "home-list-image-container-mobile" do
-      = image_tag listing.listing_images.first.image.url(:thumb), {:alt => listed_listing_title(listing), :class => "home-list-image"}
-  .home-list-details-right
-    .home-list-price
-      - if listing.price
-        .home-list-price-value
-          = humanized_money_with_symbol(listing.price).upcase
-        - price_text = nil
-        - if listing.quantity.present?
-          - price_text = t("listings.form.price.per") + " " + listing.quantity
-        - elsif listing.unit_type
-          - price_text = price_quantity_per_unit(listing)
+- cache(listing) do
+  .home-list-item
+    - if listing.listing_images.size > 0
+      = link_to listing, :class => "home-list-image-container-desktop" do
+        = image_tag listing.listing_images.first.image.url(:small_3x2), {:alt => listed_listing_title(listing), :class => "home-list-image"}
+    - if listing.listing_images.size > 0
+      = link_to listing, :class => "home-list-image-container-mobile" do
+        = image_tag listing.listing_images.first.image.url(:thumb), {:alt => listed_listing_title(listing), :class => "home-list-image"}
+    .home-list-details-right
+      .home-list-price
+        - if listing.price
+          .home-list-price-value
+            = humanized_money_with_symbol(listing.price).upcase
+          - price_text = nil
+          - if listing.quantity.present?
+            - price_text = t("listings.form.price.per") + " " + listing.quantity
+          - elsif listing.unit_type
+            - price_text = price_quantity_per_unit(listing)
 
-        - if price_text.present?
-          .home-list-price-quantity{:title => price_text}
-            = price_text
-      - else
-        .home-list-listing-shape-value
-          = shape_name(listing)
-
-  %div{:class => (listing.listing_images.size > 0 ? "home-list-details-with-image" : "")}
-    %h2.home-list-title
-      = link_to listing do
-        = listing.title
-        - if @current_community.show_category_in_listing_list
-          %a.home-share-type-link{:href => root_path(:transaction_type => shape_name_map[listing.listing_shape_id], :view => :list)}
-            = icon_tag(listing.icon_name, ["icon-fix"])
+          - if price_text.present?
+            .home-list-price-quantity{:title => price_text}
+              = price_text
+        - else
+          .home-list-listing-shape-value
             = shape_name(listing)
 
-  .home-list-author{:class => (listing.listing_images.size > 0 ? "home-list-author-with-listing-image" : "home-list-author-without-listing-image")}
-    .home-list-avatar
-      = small_avatar_thumb(listing.author)
-    .home-list-author-details
-      = link_to listing.author :class => "home-list-author-name" do
-        = listing.author.name(@current_community)
-      - if @current_community.testimonials_in_use
-        .home-list-author-reviews
-          - if listing.author.received_testimonials.size > 0
-            = icon_tag("testimonial")
-            = pluralize(listing.author.received_testimonials.size, t(".review"), t(".reviews"))
+    %div{:class => (listing.listing_images.size > 0 ? "home-list-details-with-image" : "")}
+      %h2.home-list-title
+        = link_to listing do
+          = listing.title
+          - if @current_community.show_category_in_listing_list
+            %a.home-share-type-link{:href => root_path(:transaction_type => shape_name_map[listing.listing_shape_id], :view => :list)}
+              = icon_tag(listing.icon_name, ["icon-fix"])
+              = shape_name(listing)
 
-  .home-list-price-mobile{:class => (listing.listing_images.size > 0 ? "home-list-price-mobile-with-listing-image" : "home-list-price-mobile-without-listing-image")}
-    - if listing.price
-      = humanized_money_with_symbol(listing.price).upcase
-    - else
-      %span.smaller
-        = shape_name(listing)
+    .home-list-author{:class => (listing.listing_images.size > 0 ? "home-list-author-with-listing-image" : "home-list-author-without-listing-image")}
+      .home-list-avatar
+        = small_avatar_thumb(listing.author)
+      .home-list-author-details
+        = link_to listing.author :class => "home-list-author-name" do
+          = listing.author.name(@current_community)
+        - if @current_community.testimonials_in_use
+          .home-list-author-reviews
+            - if listing.author.received_testimonials.size > 0
+              = icon_tag("testimonial")
+              = pluralize(listing.author.received_testimonials.size, t(".review"), t(".reviews"))
+
+    .home-list-price-mobile{:class => (listing.listing_images.size > 0 ? "home-list-price-mobile-with-listing-image" : "home-list-price-mobile-without-listing-image")}
+      - if listing.price
+        = humanized_money_with_symbol(listing.price).upcase
+      - else
+        %span.smaller
+          = shape_name(listing)