--- a/app/views/homepage/_list_item.haml
+++ b/app/views/homepage/_list_item.haml
@@ -1,4 +1,4 @@
-- cache(["list_item", listing, listing.author, @current_community]) do
+- cache(["list_item", listing, listing.author, @current_community, MoneyRails::Configuration.no_cents_if_whole]) do
   .home-list-item
     - if listing.listing_images.size > 0
       = link_to listing, :class => "home-list-image-container-desktop" do