This should increase front page performance pretty easily. Performance testing with siege on `testing` gave an increase of 1,08 -> 2,02 trans/sec, details below:

Before:
```
Transactions:		         194 hits
Availability:		      100.00 %
Elapsed time:		      179.16 secs
Data transferred:	        1.86 MB
Response time:		        4.03 secs
Transaction rate:	        1.08 trans/sec
Throughput:		        0.01 MB/sec
Concurrency:		        4.36
Successful transactions:         194
Failed transactions:	           0
Longest transaction:	        9.96
Shortest transaction:	        1.43
```

After:
```
Transactions:		         364 hits
Availability:		      100.00 %
Elapsed time:		      179.93 secs
Data transferred:	        3.38 MB
Response time:		        1.94 secs
Transaction rate:	        2.02 trans/sec
Throughput:		        0.02 MB/sec
Concurrency:		        3.92
Successful transactions:         364
Failed transactions:	           0
Longest transaction:	       10.77
Shortest transaction:	        0.74
```