@mporkola 2015-06-02
The cache key is now formed from `listing`, `listing.author` and `@current_community`, which should cover all invalidation cases. Performance didn't take a hit:

```
Transactions:		         399 hits
Availability:		      100.00 %
Elapsed time:		      179.68 secs
Data transferred:	        3.82 MB
Response time:		        1.74 secs
Transaction rate:	        2.22 trans/sec
Throughput:		        0.02 MB/sec
Concurrency:		        3.86
Successful transactions:         399
Failed transactions:	           0
Longest transaction:	        7.93
Shortest transaction:	        0.80
```