@mporkola 2015-06-02
Still one caveat: expiring the fragment when template is updated. `cache-digest` gem should help